pyproject.toml
[project]
name = "hexlet-code"
version = "0.1.0"
description = "page analyzer for SEO suitability"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "beautifulsoup4>=4.12.3",
    "flask>=3.1.0",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "validators>=0.34.0",
]

[dependency-groups]
dev = [
    "ruff>=0.9.2",
]
